+++ tests/tests_plugin_xls.py
@@ -65,7 +65,7 @@ class PluginXlsTestCase(utils.RowsTestMixIn, unittest.TestCase):
             expected_meta={
                 "imported_from": "xls",
                 "filename": self.filename,  # will check `source.uri`
-                "sheet_name": "Sheet1",
+                "name": "Sheet1",
             },
         )
 
@@ -78,7 +78,7 @@ class PluginXlsTestCase(utils.RowsTestMixIn, unittest.TestCase):
                 expected_meta={
                     "imported_from": "xls",
                     "filename": self.filename,
-                    "sheet_name": "Sheet1",
+                    "name": "Sheet1",
                 },
             )